dash/requirements.txt
auth0-python              ==3.14.0   
authlib
dash                      ==1.19.0
dash-bootstrap-components ==1.2.0
dash-core-components      ==1.15.0
dash-html-components      ==1.1.2
dash-renderer             ==1.9.0
dash-table                ==4.11.2
Flask                     ==1.1.2
Flask-Compress            ==1.8.0
glob2                     ==0.7
Jinja2                    ==2.11.3
numpy                     ==1.20.0
pandas                    ==1.2.1
plotly                    ==4.14.3
pydantic
python-dotenv             ==0.15.0
requests                  ==2.25.1
urllib3                   ==1.26.3
Werkzeug                  ==1.0.1
